Fixed problem with beaver not preloading on cache clearing (#5323)
