+++ tests/Unit/inc/Engine/Preload/Subscriber/onPermalinkChanged.php
@@ -6,6 +6,7 @@ use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Engine\Preload\Activation\Activation;
 use WP_Rocket\Engine\Preload\Controller\ClearCache;
 use WP_Rocket\Engine\Preload\Controller\LoadInitialSitemap;
+use WP_Rocket\Engine\Preload\Controller\Queue;
 use WP_Rocket\Engine\Preload\Database\Queries\Cache;
 use WP_Rocket\Engine\Preload\Subscriber;
 use WP_Rocket\Tests\Unit\TestCase;
@@ -21,6 +22,7 @@ class Test_OnPermalinkChanged extends TestCase
 	protected $activation;
 	protected $mobile_detect;
 	protected $clear_cache;
+	protected $queue;
 
 	protected function setUp(): void
 	{
@@ -31,7 +33,8 @@ class Test_OnPermalinkChanged extends TestCase
 		$this->activation = Mockery::mock(Activation::class);
 		$this->mobile_detect = Mockery::mock(WP_Rocket_Mobile_Detect::class);
 		$this->clear_cache = Mockery::mock(ClearCache::class);
-		$this->subscriber = new Subscriber($this->options, $this->controller, $this->query, $this->activation, $this->mobile_detect, $this->clear_cache );
+		$this->queue = Mockery::mock(Queue::class);
+		$this->subscriber = new Subscriber($this->options, $this->controller, $this->query, $this->activation, $this->mobile_detect, $this->clear_cache, $this->queue );
 	}
 
 	public function testShouldDoAsExpected() {
